Move polyswarmd server start to app for easier test imports
