+++ src/polyswarmd/views/bounties.py
@@ -11,7 +11,7 @@ from requests import HTTPError
 from polyswarmartifact import ArtifactType
 from polyswarmartifact.schema import Assertion as AssertionMetadata
 from polyswarmartifact.schema import Bounty as BountyMetadata
-from polyswarmd import app, cache
+from polyswarmd.app import app, cache
 from polyswarmd.views import eth
 from polyswarmd.services.artifact.exceptions import ArtifactException
 from polyswarmd.utils.bloom import FILTER_BITS, BloomFilter
