+++ src/polyswarmd/views/eth.py
@@ -16,7 +16,7 @@ from web3.exceptions import MismatchedABI
 from web3.module import Module
 from web3.utils.events import get_event_data
 
-from polyswarmd import cache
+from polyswarmd.app import cache
 from polyswarmd.utils.decorators.chains import chain
 from polyswarmd.utils.response import failure, success
 from polyswarmd.websockets import messages
