+++ src/polyswarmd/wsgi.py
@@ -25,7 +25,7 @@ def app(*args, **kwargs):
         logging.exception(e)
         sys.exit(-1)
 
-    from polyswarmd import app as application
+    from polyswarmd.app import app as application
 
     logger.critical("polyswarmd is ready!")
     return application
